setting background image according to theme
