+++ lib/main.dart
@@ -7,13 +7,12 @@ import 'package:flutter/services.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 
 // Files
-import './api/unsplash_api_service.dart';
 import './controllers/activity.dart';
 import './models/activity.dart';
 import './models/global_setting.dart';
 import './models/mood.dart';
 import './routes/route_generator.dart';
-import './controllers/global_setting.dart';
+import './utils/brightness.dart';
 
 /// Starting point of the application.
 void main() async {
@@ -37,22 +36,6 @@ Future<void> hiveInitialize() async {
   Hive.registerAdapter<User>(UserAdapter());
   await Hive.openBox<Activity>(activityBoxName);
   await Hive.openBox(globalSettingBoxName);
-}
-
-// function to return brightness for theme data
-Brightness get brightness {
-  var mode = GlobalSettingController.getBrightnessOption();
-  Brightness themeMode;
-  if (mode == BrightnessOption.light) {
-    themeMode = Brightness.light;
-  }
-  else if (mode == BrightnessOption.dark) {
-    themeMode = Brightness.dark;
-  }
-  else if (mode == BrightnessOption.auto) {
-    themeMode = WidgetsBinding.instance.window.platformBrightness;
-  }
-  return themeMode;
 }
 
 /// MyApp is the most Parent widget and initialises the main route.
