+++ lib/widgets/background_images.dart
@@ -7,8 +7,7 @@ import 'package:star_book/api/errors_toast.dart';
 // Files
 import '../models/unsplash_photo.dart';
 import '../api/unsplash_api_service.dart';
-import 'package:star_book/models/unsplash_photo.dart';
-import 'package:star_book/api/unsplash_api_service.dart';
+import '../utils/brightness.dart';
 
 class BackgroundImage extends StatefulWidget {
   const BackgroundImage({
@@ -59,11 +58,12 @@ class _BackgroundImageState extends State<BackgroundImage> {
               handleErrors(_err, errToast);
             }
           } else {
+            Brightness _brightness = brightness;
             child = Image.asset(
-              "new_bg.jpeg",
-              fit: BoxFit.cover,
-              width: double.infinity,
-              height: double.infinity,
+                _brightness == Brightness.light ? "backup-bg-image.JPG" : "bg_dark.jpg",
+                fit: BoxFit.cover,
+                width: double.infinity,
+                height: double.infinity,
             );
           }
           return c.AnimatedSwitcher(
