+++ pubspec.yaml
@@ -76,6 +76,7 @@ flutter:
     - new_bg.jpeg
     - backup-bg-image.JPG
     - icon.png
+    - bg_dark.jpg
     # - .env
     # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
